Refactor(client): 커뮤니티 페이지/상세 페이지 로직 분리 (#363)
* fix: 게시물 상세 컴포넌트 이름 변경

* refactor: 커뮤니티 페이지 게시글 목록 FeedList 컴포넌트로 분리

* fix: navigate 핸들링 방식 변경

* refactor: modal 로직 분리

* chore: 게시물 상세 상단부 파일명 변경

* chore: 댓글 입력창 컴포넌트명 변경

* refactor: 댓글 목록 컴포넌트 UserCommentList로 분리 및 스타일 추가

* refactor: 삭제 모달 로직 통합 및 입력 제한 상수화

* refactor: CommunityDetail 페이지에서 불필요한 로직 제거 및 FeedContent 컴포넌트로 분리

* feat: 커뮤니티 상세 페이지 로직 분리

* refactor: useNavigateTo 훅을 추가하여 내비게이션 로직 간소화

* refactor: useNavigateTo 훅을 사용하여 내비게이션 로직 간소화

* feat: CommunityBody 컴포넌트명 변경

* refactor: CommunityModal에서 버튼 상태 관리 로직 개선
