+++ apps/client/src/widgets/community/components/comment-input-box/comment-input-box.css.ts
