+++ apps/client/src/widgets/community/components/feed-content/feed-content.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  display: 'flex',
+  flexDirection: 'column',
+  padding: '2.4rem 1.6rem 0 1.6rem',
+  gap: '1.6rem',
+});
