+++ apps/client/src/widgets/community/components/feed-detail-info/feed-detail-info.css.ts
