+++ apps/client/src/widgets/community/components/feed-list-item/feed-list-item.css.ts
