+++ apps/client/src/widgets/community/constant/modal-delete-content.ts
@@ -1,0 +1,10 @@
+export const DELETE_MODAL = {
+  FEED: {
+    title: '이 글을 삭제할까요?',
+    content: '삭제한 글/댓글은 복원되지 않습니다.',
+  },
+  COMMENT: {
+    title: '이 댓글을 삭제할까요?',
+    content: '삭제한 댓글은 복원되지 않습니다.',
+  },
+};
