+++ apps/client/src/widgets/community/types/community-modal.type.ts
@@ -1,0 +1,1 @@
+export type ModalType = 'feed' | 'comment';
